requirements.txt
InquirerPy==0.3.4
pathlib==1.0.1
typing==3.7.4.3
typer==0.12.3
doc2md==0.1.0
rich==13.7.1
python-docx>=0.8.11
markdown>=3.5.1
beautifulsoup4>=4.12.2
tinycss>=0.4
